others/requirements-spark.txt
pyspark
numpy
union
flytekitplugins-spark>=1.13.1a2
# comment
plotly
flytekit>=1.13.1a2
wget